pyproject.toml
[tool.poetry]
name = "RAG GPT"
version = "0.1.0"
description = "A retrieval augmented generation GPT"
authors = ['Hanif Shah']



[tool.poetry.dependencies]
python= ">3.10.0,<3.12"
langchain = ">=0.2.13,<0.3.0"
crewai= "0.51.1"
langchain-community = "0.2.12"
python-dotenv= "1.0.0"  
langchain-together = "^0.1.5"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.32"
pymysql = "^1.1.1"
mysql-connector-python = "^9.0.0"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"